+++ src/alignment/justified.rs
@@ -1,9 +1,9 @@
 //! Fully justified text.
 use crate::{
     alignment::{HorizontalTextAlignment, VerticalTextAlignment},
-    parser::{Parser, Token},
+    parser::Token,
     rendering::{
-        cursor::Cursor, line::StyledLinePixelIterator, space_config::SpaceConfig, StateFactory,
+        line::StyledLinePixelIterator, space_config::SpaceConfig, RendererFactory,
         StyledTextBoxIterator,
     },
     style::height_mode::HeightMode,
@@ -11,7 +11,7 @@ use crate::{
     StyledTextBox,
 };
 use core::marker::PhantomData;
-use embedded_graphics::{drawable::Pixel, fonts::Font, pixelcolor::PixelColor};
+use embedded_graphics::{fonts::Font, pixelcolor::PixelColor};
 
 /// Marks text to be rendered fully justified.
 #[derive(Copy, Clone, Debug)]
@@ -76,94 +76,37 @@ impl<F: Font + Copy> SpaceConfig for JustifiedSpaceConfig<F> {
     }
 }
 
-/// State variable used by the fully justified text renderer.
-#[derive(Debug)]
-pub enum State<'a, C, F>
-where
-    C: PixelColor,
-    F: Font + Copy,
-{
-    /// Starts processing a line.
-    NextLine(Option<Token<'a>>, Cursor<F>, Parser<'a>),
-
-    /// Renders the processed line.
-    DrawLine(StyledLinePixelIterator<'a, C, F, JustifiedSpaceConfig<F>, Justified>),
-}
-
-impl<'a, C, F, V, H> StateFactory<'a, F> for StyledTextBox<'a, C, F, Justified, V, H>
+impl<'a, C, F, V, H> RendererFactory<'a, C> for StyledTextBox<'a, C, F, Justified, V, H>
 where
     C: PixelColor,
     F: Font + Copy,
     V: VerticalTextAlignment,
     H: HeightMode,
 {
-    type PixelIteratorState = State<'a, C, F>;
+    type Renderer = StyledTextBoxIterator<'a, C, F, Justified, V, H, JustifiedSpaceConfig<F>>;
 
     #[inline]
     #[must_use]
-    fn create_state(&self, cursor: Cursor<F>, parser: Parser<'a>) -> Self::PixelIteratorState {
-        State::NextLine(None, cursor, parser)
-    }
-}
+    fn create_renderer(&self) -> Self::Renderer {
+        StyledTextBoxIterator::new(self, |style, carried, cursor, parser| {
+            let max_line_width = cursor.line_width();
+            let (width, total_whitespace_count, t) =
+                style.measure_line(&mut parser.clone(), carried.clone(), max_line_width);
 
-impl<C, F, V, H> Iterator for StyledTextBoxIterator<'_, C, F, Justified, V, H>
-where
-    C: PixelColor,
-    F: Font + Copy,
-    V: VerticalTextAlignment,
-    H: HeightMode,
-{
-    type Item = Pixel<C>;
+            let space =
+                max_line_width - (width - total_whitespace_count * F::total_char_width(' '));
+            let stretch_line = t.is_some() && t != Some(Token::NewLine);
 
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            match self.state {
-                State::NextLine(ref carried_token, cursor, ref mut parser) => {
-                    if carried_token.is_none() && parser.is_empty() {
-                        break None;
-                    }
+            let space_info = if stretch_line && total_whitespace_count != 0 {
+                let space_width = space / total_whitespace_count;
+                let extra_pixels = space % total_whitespace_count;
+                JustifiedSpaceConfig::new(space_width, extra_pixels)
+            } else {
+                JustifiedSpaceConfig::default()
+            };
 
-                    let parser_clone = parser.clone();
-                    let max_line_width = cursor.line_width();
-                    let (width, total_whitespace_count, t) =
-                        self.style
-                            .measure_line(parser, carried_token.clone(), max_line_width);
-
-                    let space = max_line_width
-                        - (width - total_whitespace_count * F::total_char_width(' '));
-                    let stretch_line = t.is_some() && t != Some(Token::NewLine);
-
-                    let space_info = if stretch_line && total_whitespace_count != 0 {
-                        let space_width = space / total_whitespace_count;
-                        let extra_pixels = space % total_whitespace_count;
-                        JustifiedSpaceConfig::new(space_width, extra_pixels)
-                    } else {
-                        JustifiedSpaceConfig::default()
-                    };
-
-                    self.state = State::DrawLine(StyledLinePixelIterator::new(
-                        parser_clone,
-                        cursor,
-                        space_info,
-                        self.style,
-                        carried_token.clone(),
-                    ));
-                }
-
-                State::DrawLine(ref mut line_iterator) => {
-                    if let pixel @ Some(_) = line_iterator.next() {
-                        break pixel;
-                    }
-
-                    self.state = State::NextLine(
-                        line_iterator.remaining_token(),
-                        line_iterator.cursor,
-                        line_iterator.parser(),
-                    );
-                }
-            }
-        }
+            StyledLinePixelIterator::new(parser, cursor, space_info, style, carried)
+        })
     }
 }