+++ src/alignment/right.rs
@@ -1,15 +1,14 @@
 //! Right aligned text.
 use crate::{
     alignment::{HorizontalTextAlignment, VerticalTextAlignment},
-    parser::{Parser, Token},
     rendering::{
-        cursor::Cursor, line::StyledLinePixelIterator, space_config::UniformSpaceConfig,
-        StateFactory, StyledTextBoxIterator,
+        line::StyledLinePixelIterator, space_config::UniformSpaceConfig, RendererFactory,
+        StyledTextBoxIterator,
     },
     style::height_mode::HeightMode,
     StyledTextBox,
 };
-use embedded_graphics::{drawable::Pixel, fonts::Font, pixelcolor::PixelColor};
+use embedded_graphics::{fonts::Font, pixelcolor::PixelColor};
 
 /// Marks text to be rendered right aligned.
 #[derive(Copy, Clone, Debug)]
@@ -19,83 +18,32 @@ impl HorizontalTextAlignment for RightAligned {
     const ENDING_SPACES: bool = false;
 }
 
-/// State variable used by the right aligned text renderer.
-#[derive(Debug)]
-pub enum State<'a, C, F>
-where
-    C: PixelColor,
-    F: Font + Copy,
-{
-    /// Starts processing a line.
-    NextLine(Option<Token<'a>>, Cursor<F>, Parser<'a>),
-
-    /// Renders the processed line.
-    DrawLine(StyledLinePixelIterator<'a, C, F, UniformSpaceConfig<F>, RightAligned>),
-}
-
-impl<'a, C, F, V, H> StateFactory<'a, F> for StyledTextBox<'a, C, F, RightAligned, V, H>
+impl<'a, C, F, V, H> RendererFactory<'a, C> for StyledTextBox<'a, C, F, RightAligned, V, H>
 where
     C: PixelColor,
     F: Font + Copy,
     V: VerticalTextAlignment,
     H: HeightMode,
 {
-    type PixelIteratorState = State<'a, C, F>;
+    type Renderer = StyledTextBoxIterator<'a, C, F, RightAligned, V, H, UniformSpaceConfig<F>>;
 
     #[inline]
     #[must_use]
-    fn create_state(&self, cursor: Cursor<F>, parser: Parser<'a>) -> Self::PixelIteratorState {
-        State::NextLine(None, cursor, parser)
-    }
-}
+    fn create_renderer(&self) -> Self::Renderer {
+        StyledTextBoxIterator::new(self, |style, carried, mut cursor, parser| {
+            let max_line_width = cursor.line_width();
+            let (width, _, _) =
+                style.measure_line(&mut parser.clone(), carried.clone(), max_line_width);
+            cursor.advance_unchecked(max_line_width - width);
 
-impl<C, F, V, H> Iterator for StyledTextBoxIterator<'_, C, F, RightAligned, V, H>
-where
-    C: PixelColor,
-    F: Font + Copy,
-    V: VerticalTextAlignment,
-    H: HeightMode,
-{
-    type Item = Pixel<C>;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            match self.state {
-                State::NextLine(ref carried_token, mut cursor, ref mut parser) => {
-                    if carried_token.is_none() && parser.is_empty() {
-                        break None;
-                    }
-
-                    let parser_clone = parser.clone();
-                    let max_line_width = cursor.line_width();
-                    let (width, _, _) =
-                        self.style
-                            .measure_line(parser, carried_token.clone(), max_line_width);
-                    cursor.advance_unchecked(max_line_width - width);
-
-                    self.state = State::DrawLine(StyledLinePixelIterator::new(
-                        parser_clone,
-                        cursor,
-                        UniformSpaceConfig::default(),
-                        self.style,
-                        carried_token.clone(),
-                    ));
-                }
-
-                State::DrawLine(ref mut line_iterator) => {
-                    if let pixel @ Some(_) = line_iterator.next() {
-                        break pixel;
-                    }
-
-                    self.state = State::NextLine(
-                        line_iterator.remaining_token(),
-                        line_iterator.cursor,
-                        line_iterator.parser(),
-                    );
-                }
-            }
-        }
+            StyledLinePixelIterator::new(
+                parser,
+                cursor,
+                UniformSpaceConfig::default(),
+                style,
+                carried,
+            )
+        })
     }
 }