+++ src/rendering/character.rs
@@ -1,13 +1,12 @@
 //! Character rendering.
-use core::marker::PhantomData;
-use core::ops::Range;
+use core::{marker::PhantomData, ops::Range};
 use embedded_graphics::{prelude::*, style::TextStyle};
 
 /// Represents a glyph (a symbol) to be drawn.
 #[derive(Copy, Clone, Debug)]
 pub struct Glyph<F: Font> {
     _font: PhantomData<F>,
-    char_idx: u32,
+    char_glyph_offset: u32,
 }
 
 impl<F> Glyph<F>
@@ -18,7 +17,7 @@ where
     #[inline]
     #[must_use]
     pub fn new(c: char) -> Self {
-        let char_offset = F::char_offset(if c == '\u{A0}' { ' ' } else { c });
+        let char_offset = F::char_offset(c);
         let char_per_row = F::FONT_IMAGE_WIDTH / F::CHARACTER_SIZE.width;
 
         // Top left corner of character, in pixels.
@@ -27,7 +26,7 @@ where
 
         Self {
             _font: PhantomData,
-            char_idx: char_x + char_y * F::FONT_IMAGE_WIDTH,
+            char_glyph_offset: char_x + char_y * F::FONT_IMAGE_WIDTH,
         }
     }
 
@@ -42,7 +41,8 @@ where
         // + Character row offset (row 0 = 0, row 1 = (192 * 8) = 1536)
         // + X offset for the pixel block that comprises this char
         // + Y offset for pixel block
-        let bitmap_bit_index = self.char_idx + p.x as u32 + p.y as u32 * F::FONT_IMAGE_WIDTH;
+        let bitmap_bit_index =
+            self.char_glyph_offset + p.x as u32 + p.y as u32 * F::FONT_IMAGE_WIDTH;
 
         let bitmap_byte = bitmap_bit_index / 8;
         let bitmap_bit = bitmap_bit_index % 8;