+++ src/rendering/line.rs
@@ -10,14 +10,13 @@ use crate::{
         whitespace::EmptySpaceIterator,
     },
     style::{height_mode::HeightMode, TextBoxStyle},
-    utils::font_ext::FontExt,
 };
 use core::ops::Range;
 use embedded_graphics::{prelude::*, style::TextStyle};
 
 /// Internal state used to render a line.
 #[derive(Debug)]
-pub enum State<C, F>
+enum State<C, F>
 where
     C: PixelColor,
     F: Font + Copy,
@@ -30,9 +29,6 @@ where
 
     /// Render a block of whitespace.
     Space(EmptySpaceIterator<C, F>),
-
-    /// Signal that the renderer has finished.
-    Done,
 }
 
 /// Pixel iterator to render a single line of styled text.
@@ -44,9 +40,6 @@ where
     SP: SpaceConfig<Font = F>,
     A: HorizontalTextAlignment,
 {
-    /// Position information.
-    pub cursor: Cursor<F>,
-
     state: State<C, F>,
     style: TextStyle<C, F>,
     display_range: Range<i32>,
@@ -75,7 +68,6 @@ where
         H: HeightMode,
     {
         Self {
-            cursor,
             state: State::FetchNext,
             style: style.text_style,
             display_range: H::calculate_displayed_row_range(&cursor),
@@ -100,6 +92,13 @@ where
         self.inner.parser.clone()
     }
 
+    /// When finished, this method returns the cursor object.
+    #[must_use]
+    #[inline]
+    pub fn cursor(&self) -> Cursor<F> {
+        self.inner.cursor
+    }
+
     fn is_anything_displayed(&self) -> bool {
         self.display_range.start < self.display_range.end
     }
@@ -120,38 +119,33 @@ where
             match self.state {
                 // No token being processed, get next one
                 State::FetchNext => {
-                    self.state = match self.inner.next() {
+                    match self.inner.next() {
                         Some(RenderElement::PrintedCharacter(c)) => {
-                            let char_width = F::total_char_width(c);
-                            let pos = self.cursor.position;
-                            self.cursor.advance_unchecked(char_width);
-
                             if self.is_anything_displayed() {
-                                State::Char(StyledCharacterIterator::new(
+                                self.state = State::Char(StyledCharacterIterator::new(
                                     c,
-                                    pos,
+                                    self.inner.pos,
                                     self.style,
                                     self.display_range.clone(),
-                                ))
+                                ));
                             } else {
-                                State::FetchNext
+                                self.state = State::FetchNext;
                             }
                         }
 
                         Some(RenderElement::Space(space_width, _)) => {
-                            let pos = self.cursor.position;
-                            self.cursor.advance_unchecked(space_width);
                             if self.is_anything_displayed() {
-                                State::Space(EmptySpaceIterator::new(space_width, pos, self.style))
+                                self.state = State::Space(EmptySpaceIterator::new(
+                                    space_width,
+                                    self.inner.pos,
+                                    self.style,
+                                ));
                             } else {
-                                State::FetchNext
+                                self.state = State::FetchNext;
                             }
                         }
 
-                        None => {
-                            self.cursor = self.inner.cursor;
-                            State::Done
-                        }
+                        None => break None,
                     };
                 }
 
@@ -169,10 +163,6 @@ where
                     }
 
                     self.state = State::FetchNext;
-                }
-
-                State::Done => {
-                    break None;
                 }
             }
         }